# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation of afroluxe.",
        "title": "Afroluxe API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "afroluxe.heroku.app",
    "basePath": "/",
    "paths": {
...
        "/stylist": {
            "post": {
                "description": "Creates new stylist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Stylist"
                ],
                "summary": "Creates new stylist",
                "parameters": [
                    {
                        "description": "Stylist",
                        "name": "stylist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Stylist"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    }
                }
            }
        },
        "/stylist/images": {
            "post": {
                "description": "Reviews stylist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Stylist"
                ],
                "summary": "Reviews stylist",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Stylist Images",
                        "name": "images",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    }
                }
            }
        },
        "/stylist/review": {
            "post": {
                "description": "Reviews stylist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Stylist"
                ],
                "summary": "Reviews stylist",
                "parameters": [
                    {
                        "description": "Stylist",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Review"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    }
                }
            }
        },
        "/stylist/{id}": {
            "get": {
                "description": "Verifies user email",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Stylist"
                ],
                "summary": "Gets stylist based on stylist id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stylist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.BusinessHour": {
            "type": "object",
            "required": [
                "friday",
                "monday",
                "saturday",
                "sunday",
                "thursday",
                "tuesday",
                "wednesday"
            ],
            "properties": {
                "friday": {
                    "type": "string"
                },
                "monday": {
                    "type": "string"
                },
                "saturday": {
                    "type": "string"
                },
                "sunday": {
                    "type": "string"
                },
                "thursday": {
                    "type": "string"
                },
                "tuesday": {
                    "type": "string"
                },
                "wednesday": {
                    "type": "string"
                }
            }
        },
        "models.Image": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.Location": {
            "type": "object",
            "required": [
                "latitude",
                "longitude"
            ],
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
...
        "models.Review": {
            "type": "object",
            "required": [
                "body",
                "pricing_rating",
                "recommendation_rating",
                "service_rating",
                "stylist_id",
                "user_id"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "pricing_rating": {
                    "type": "integer"
                },
                "recommendation_rating": {
                    "type": "integer"
                },
                "service_rating": {
                    "type": "integer"
                },
                "stylist_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.Service": {
            "type": "object",
            "required": [
                "currency_name",
                "currency_symbol",
                "name",
                "price"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "currency_name": {
                    "type": "string"
                },
                "currency_symbol": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "stylist_id": {
                    "type": "string"
                }
            }
        },
        "models.Stylist": {
            "type": "object",
            "required": [
                "bio",
                "business_hour",
                "business_name",
                "business_type",
                "category",
                "home_service",
                "location",
                "services",
                "size",
                "user_id"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "business_hour": {
                    "$ref": "#/definitions/models.BusinessHour"
                },
                "business_name": {
                    "type": "string"
                },
                "business_type": {
                    "type": "string"
                },
                "category": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "type": "integer"
                },
                "home_service": {
                    "type": "boolean"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Image"
                    }
                },
                "location": {
                    "$ref": "#/definitions/models.Location"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Service"
                    }
                },
                "size": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}